[ao] Add method in ModelReport to generate visualizer
Summary: We created a ModelReportVisualizer class, and the primary
way it is envisioned that it is accessed is:

```
model_report_visualizer = model_reporter.generate_visualizer()
```

This method only works after reports have been generated and it takes in
the generated reports and reformats them to be ordered by module, into
the format required by the ModelReportVisualization. It then generates
the visualizer instance and returns that to the user.

Test Plan: python test/test_quantization.py TestFxModelReportClass.test_generate_visualizer

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 7f4eeacee1947f63ae49f76452260643517ba7d4
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81589