[ao] Get feature and module names from ModelReportVisualizer class
Summary: Added the functionality to be able to get the feature names and
module_fqns from the ModelReportVisualizer class. The purpose of this
addition is so that users can see the exact set of module_fqns or
feature names that they can filter based on, and use this information to
perform their filtering.

Test Plan: python test/test_quantization.py
TestFxModelReportVisualizer.test_get_modules_and_features

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: acbf0da4d2c1d14c1988ec1dbff4643ed02ff9bb
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81647